AGENT/requirements.txt
fastapi
uvicorn
pymysql
python-dotenv
langgraph
langchain-openai
